+++ pkg/release/release_notes.go
@@ -6,7 +6,7 @@ import (
 	"text/template"
 
 	"github.com/giantswarm/microerror"
-	"github.com/giantswarm/release-operator/v4/api/v1alpha1"
+	"github.com/giantswarm/releases/sdk/api/v1alpha1"
 
 	"github.com/giantswarm/devctl/v7/pkg/release/changelog"
 )